Long file support in buffer layer
A giant (but sparse) cache file is created for Rezzed resource forks,
allowing for 17 MB each. This exceeds 2 GB when the 121st file is
opened, requiring 64-bit file offsets.

+++ multifork-3.c
@@ -424,7 +424,7 @@ static uint32_t readin(uint32_t rezfid, uint32_t scratchfid, uint64_t offset) {
 		int32_t rezoffset = rbufseek;
 
 		int32_t bodylen;
-		int32_t fixup = wbufseek;
+		int64_t fixup = wbufseek;
 		for (int i=0; i<4; i++) Write(0);
 		bodylen = RezBody();
 		err = Overwrite(&bodylen, fixup, 4);
